[#98] Add generics to the KeycloakContainer class
