+++ README.md
@@ -189,6 +189,29 @@ KeycloakContainer keycloak = new KeycloakContainer()
 
 You have to provide a list of resolvable `File`s.
 
+### Extending KeycloakContainer
+
+In case you need a custom implementation of the default `KeycloakContainer`, you should inherit from `ExtendableKeycloakContainer`. This allows to set the generics and use your custom implementation without the need for type casts.  
+
+```java
+public class MyCustomKeycloakContainer extends ExtendableKeycloakContainer<MyCustomKeycloakContainer> {
+
+	public MyCustomKeycloakContainer() {
+		super();
+	}
+
+	public MyCustomKeycloakContainer(String dockerImageName) {
+		super(dockerImageName);
+	}
+	
+}
+
+...
+
+MyCustomKeycloakContainer keycloakContainer = new MyCustomKeycloakContainer()
+    .withAdminPassword("password");
+```
+
 #### TIPP
 
 If you want/need to use dependencies from e.g. Maven (or Gradle), you can use [ShrinkWrap Resolvers](https://github.com/shrinkwrap/resolver).
